# openapi.yaml
openapi: 3.0.3
info:
  title: NavalhIA API
  description: API para gestão no NavalhIA (agendamentos, clientes, barbeiros, serviços). Autenticação via JWT (Bearer) ou API Key (header X-API-Key).
  version: 1.0.0
servers:
  - url: https://api.navalhia.com.br
    description: API (substitua pela URL real)
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Token e perfil
        "401":
          description: Credenciais inválidas
  /api/auth/me:
    get:
      summary: Perfil do usuário autenticado
      tags: [Auth]
      responses:
        "200":
          description: Perfil (id, email, barbershop_id, role, must_change_password)
        "401":
          description: Não autorizado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key